fix: removed unused code for querycontext
